requirements.txt
Django==3.2.6
django-filter==2.4.0
djangorestframework==3.12.4
Markdown==3.3.4
Pillow==8.3.1
django-allauth==0.45.0
django-rest-auth==0.9.5
mysqlclient